25 Jan 2nd Shift
This is for the finance manager module and the milk production.
I have removed the cash flow statement from the fince module tab on the side bar and put in the Mega menu in the top nav bar.

In the views, Osca had used the $deleten and $updaten variables but i have failed to see where they originate from. The one for permissions.